Move IStatisticDefinition to arcgis-rest-types
