+++ packages/arcgis-rest-types/src/index.ts
@@ -6,3 +6,4 @@ export * from "./service";
 export * from "./symbol";
 export * from "./user";
 export * from "./webmap";
+export * from "./statisticDefinition";
